toolkit/components/extensions/test/mochitest/test_ext_proxy.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for the proxy API</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
/* eslint no-unused-vars: ["error", {"args": "none", "varsIgnorePattern": "^(FindProxyForURL)$"}] */

"use strict";

async function testProxyScript(script, expected) {
  let extension = ExtensionTestUtils.loadExtension({
    background() {
      let errorReceived = false;
      browser.proxy.onProxyError.addListener(error => {
        if (!errorReceived) {
          errorReceived = true;
          browser.test.sendMessage("proxy-error-received", error);
        }
      });

      browser.proxy.register("proxy_script.js").then(() => {
        browser.test.sendMessage("ready");
      });

      browser.test.onMessage.addListener(msg => {
        if (msg === "unregister-proxy-script") {
          browser.proxy.unregister().then(() => {
            browser.test.notifyPass("proxy");
          });
        }
      });
    },
    manifest: {
      "permissions": ["proxy"],
    },
    files: {
      "proxy_script.js": String(script).replace(/^.*?\{([^]*)\}$/, "$1"),
    },
  });

  await extension.startup();
  await extension.awaitMessage("ready");

  let win = window.open("http://example.com/");
  let error = await extension.awaitMessage("proxy-error-received");
  is(error.message, expected.message, "Correct error message received");

  if (expected.errorInfo) {
    ok(error.fileName.includes("proxy_script.js"), "Error should include file name");
    is(error.lineNumber, 3, "Error should include line number");
    ok(error.stack.includes("proxy_script.js:3:9"), "Error should include stack trace");
  }

  extension.sendMessage("unregister-proxy-script");
  await extension.awaitFinish("proxy");

  win.close();
  await extension.unload();
}

add_task(async function test_invalid_FindProxyForURL_type() {
  await testProxyScript(
    () => { }, {
      message: "The proxy script must define FindProxyForURL as a function",
    });

  await testProxyScript(
    () => {
      var FindProxyForURL = 5; // eslint-disable-line mozilla/var-only-at-top-level
    }, {
      message: "The proxy script must define FindProxyForURL as a function",
    });
});

add_task(async function test_invalid_FindProxyForURL_return_types() {
  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return 5;
      }
    }, {
      message: "FindProxyForURL: Return type must be a string",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "INVALID";
      }
    }, {
      message: "FindProxyForURL: Unrecognized proxy type: \"INVALID\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "SOCKS";
      }
    }, {
      message: "FindProxyForURL: Missing argument for proxy type: \"SOCKS\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "PROXY 1.2.3.4:8080 EXTRA";
      }
    }, {
      message: "FindProxyForURL: Too many arguments passed for proxy rule: \"PROXY 1.2.3.4:8080 EXTRA\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "PROXY :";
      }
    }, {
      message: "FindProxyForURL: Unable to parse host and port from proxy rule: \"PROXY :\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "PROXY :8080";
      }
    }, {
      message: "FindProxyForURL: Unable to parse host and port from proxy rule: \"PROXY :8080\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "PROXY ::";
      }
    }, {
      message: "FindProxyForURL: Unable to parse host and port from proxy rule: \"PROXY ::\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "PROXY 1.2.3.4:";
      }
    }, {
      message: "FindProxyForURL: Unable to parse host and port from proxy rule: \"PROXY 1.2.3.4:\"",
    });

  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return "DIRECT 1.2.3.4:8080";
      }
    }, {
      message: "FindProxyForURL: Invalid argument for proxy type: \"DIRECT\"",
    });
});

add_task(async function test_runtime_error_in_FindProxyForURL() {
  await testProxyScript(
    () => {
      function FindProxyForURL() {
        return not_defined; // eslint-disable-line no-undef
      }
    }, {
      message: "not_defined is not defined",
      errorInfo: true,
    });
});

</script>

</body>
</html>
